+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Show cache hit ratio decimal points in Admin Bar node
 - Obscure secrets when displaying `WP_REDIS_SERVERS`
 - Improved CloudLinux's Accelerate WP compatibility
+- Flush Redis database when enabling object cache
+- Show a custom error message when Redis is unreachable
+- Don't allow object cache to be enabled when Redis is unreachable
 
 ## 2.3.0